+++ src/project/src/model/TransferMessage.java
@@ -4,9 +4,9 @@ public class TransferMessage {
     private final int transferId;
     private final int fromAccountId;
     private final int toAccountId;
-    private final double amount;
+    private final long amount;
 
-    public TransferMessage(int transferId, int fromAccountId, int toAccountId, double amount) {
+    public TransferMessage(int transferId, int fromAccountId, int toAccountId, long amount) {
         this.transferId = transferId;
         this.fromAccountId = fromAccountId;
         this.toAccountId = toAccountId;
@@ -25,7 +25,7 @@ public class TransferMessage {
         return toAccountId;
     }
 
-    public double getAmount() {
+    public long getAmount() {
         return amount;
     }
 }
